requirements.txt
python-telegram-bot==12.7
pyyaml==5.4.1